Refactor investment details form middleware into it's own handler
